AAGSB-1390 fix publish rate and advise users to only use for single camera streamers

+++ include/web_video_server/web_video_server.h
@@ -59,7 +59,7 @@ protected:
   rclcpp::Node::SharedPtr nh_;
   rclcpp::WallTimer<rclcpp::VoidCallbackType>::SharedPtr cleanup_timer_;
   int ros_threads_;
-  double publish_rate_;
+  int publish_rate_;
   std::vector<std::string> advertise_topics_;
   int port_;
   std::string address_;
